// src/modules/database-driver/connector/grpc/gen/GrpcEnums_pb.ts
export enum GrpcTaskSimplifiedState {
  /**
   * *
   * Task is waiting in the queue to be executed.
   *
   * @generated from enum value: TASK_QUEUED = 0;
   */
  TASK_QUEUED = 0,

  /**
   * *
   * Task is currently running.
   *
   * @generated from enum value: TASK_RUNNING = 1;
   */
  TASK_RUNNING = 1,

  /**
   * *
   * Task has finished successfully.
   *
   * @generated from enum value: TASK_FINISHED = 2;
   */
  TASK_FINISHED = 2,

  /**
   * *
   * Task has failed.
   *
   * @generated from enum value: TASK_FAILED = 3;
   */
  TASK_FAILED = 3,

  /**
   * *
   * Task is waiting for precondition to be satisfied.
   *
   * @generated from enum value: TASK_WAITING_FOR_PRECONDITION = 4;
   */
  TASK_WAITING_FOR_PRECONDITION = 4,
}